Remove unnecessary cast which was causing a compilation warning with JDK1.6.

+++ opendj-sdk/opends/src/server/org/opends/server/protocols/ldap/LDAPConnectionHandler.java
@@ -484,7 +484,7 @@ public final class LDAPConnectionHandler extends
    *         allowed by this connection handler.
    */
   public int getMaxRequestSize() {
-    return (int) ((long) currentConfig.getMaxRequestSize());
+    return (int) currentConfig.getMaxRequestSize();
   }
